Cleans up helpers API.
