+++ helpers/__init__.py
@@ -1,9 +1,12 @@
 from helpers import template_helpers
-from helpers import application_helper
+from helpers import string_helpers
 
 
 # This will be used by Flask's @app.context_processor (in controllers.__init__)
 api = dict(
-    active_tab_class    = template_helpers.active_tab_class,
-    at                  = application_helper.at
+    active_tab_class        = template_helpers.active_tab_class,
+    at                      = template_helpers.at,
+    filter_flash_messages   = template_helpers.filter_flash_messages,
+    parameterize            = string_helpers.parameterize,
+    pluralize               = string_helpers.pluralize
 )
